Modify README.md file, add LICENCE file
